requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
bcrypt==4.2.1
PyMySQL==1.1.1
cryptography==42.0.8
python-dotenv==1.0.1
Werkzeug==3.1.3
pandas==2.2.3
openpyxl==3.1.5
xlsxwriter==3.2.0
reportlab==4.2.5
